ryu/ai_forecasting/requirements.txt
psycopg2-binary
pandas
sqlalchemy
prophet
scikit-learn